+++ src/main/java/net/silentchaos512/gear/item/blueprint/AbstractBlueprintItem.java
@@ -1,16 +1,18 @@
 package net.silentchaos512.gear.item.blueprint;
 
-import net.minecraft.world.item.Item;
-import net.minecraft.world.item.CreativeModeTab;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Rarity;
-import net.minecraft.tags.Tag;
+import net.minecraft.ChatFormatting;
 import net.minecraft.core.NonNullList;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.tags.Tag;
+import net.minecraft.world.item.*;
+import net.minecraft.world.level.Level;
+import net.minecraftforge.fml.ModList;
 import net.silentchaos512.gear.config.Config;
+import net.silentchaos512.gear.util.TextUtil;
 
-import net.minecraft.world.item.Item.Properties;
+import javax.annotation.Nullable;
+import java.util.List;
 
 public abstract class AbstractBlueprintItem extends Item implements IBlueprint {
     final boolean singleUse;
@@ -66,5 +68,28 @@ public abstract class AbstractBlueprintItem extends Item implements IBlueprint {
 
     public boolean hasStandardModel() {
         return true;
+    }
+
+    @Override
+    public void appendHoverText(ItemStack stack, @Nullable Level level, List<Component> tooltip, TooltipFlag flags) {
+        if (!isDisabled()) {
+            if (Config.Client.showJeiHints.get()) {
+                // JEI hints
+                if (ModList.get().isLoaded("jei")) {
+                    // JEI is installed! Provide a tip for noobs.
+                    tooltip.add(TextUtil.translate("item", "blueprint.jeiTip").withStyle(ChatFormatting.AQUA));
+                } else {
+                    // JEI is not installed? Educate the ultra noobs.
+                    tooltip.add(TextUtil.misc("jeiNotInstalled").withStyle(ChatFormatting.DARK_RED));
+                }
+            }
+
+            // Legacy material mixing allowed?
+            if (Config.Common.allowLegacyMaterialMixing.get()) {
+                tooltip.add(TextUtil.translate("item", "blueprint.mixing.enabled").withStyle(ChatFormatting.GREEN));
+            } else {
+                tooltip.add(TextUtil.translate("item", "blueprint.mixing.disabled").withStyle(ChatFormatting.RED));
+            }
+        }
     }
 }