+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Some tips regarding JEI to blueprints
+
 ## [2.7.6] - 2021-11-07
 ### Added
 - New config options for material and part tooltips, found in the client config (Rimevel) [#421]